requirements.txt
requests
mysql-connector-python
pycryptodome
tuya-connector-python
python-dotenv
tuya-iot-py-sdk
influxdb-client